SSTTP-551 Reformatted models to group them together more effectively
